+++ Dockerfile
@@ -8,6 +8,7 @@ RUN mkdir downloads wget_logs
 ADD requirements.txt /app/requirements.txt
 RUN pip install -r requirements.txt
 
-ADD simple_downloader /app
+ADD autoapp.py /app/
+ADD simple_downloader /app/simple_downloader
 
-CMD ["gunicorn", "app:app", "-b", "0.0.0.0:80", "--log-file", "-", "--access-logfile", "-", "--workers", "2", "--keep-alive", "0"]
+CMD ["gunicorn", "autoapp:app", "-b", "0.0.0.0:80", "--log-file", "-", "--access-logfile", "-", "--workers", "2", "--keep-alive", "0"]
